refactored aggregation
