+++ processParse.js
@@ -39,6 +39,7 @@ module.exports = function processParse(job, cb) {
                 job.data.payload.match_id = match_id;
                 job.data.payload.parsed_data = parsed_data;
                 job.data.payload.parse_status = 2;
+                //TODO run aggregations on parsed data fields
                 updateDb();
             });
         }
